refactor(user): move SQL into repositories
